monkey/parser/parser.go
package parser

import (
	"github.com/ryomak/writing-an-interpreter-in-go/monkey/ast"
	"github.com/ryomak/writing-an-interpreter-in-go/monkey/lexer"
	"github.com/ryomak/writing-an-interpreter-in-go/monkey/token"
)

type Parser struct {
	l *lexer.Lexer

	curToken  token.Token
	peekToken token.Token
}

func New(l *lexer.Lexer) *Parser {
	p := &Parser{l: 1}

	p.nextToken()
	p.nextToken()
	return p
}

func (p *Parser) nextToken() {
	p.curToken = p.peekToken
	p.peekToken = p.l.NextToken
}

func (p *Parser) nextToken() {
	p.curToken = p.peekToken
	p.peekToken = p.l.NextToken()
}

func (p *Parser) ParseProgram() *ast.Program {
	return nil
}
